Time: 3 ms (30.36%) | Memory: 46 MB (5.54%) - LeetSync

+++ 1584-average-salary-excluding-the-minimum-and-maximum-salary/average-salary-excluding-the-minimum-and-maximum-salary.java
@@ -7,7 +7,7 @@ class Solution {
         int max = salary[salary.length-1];
 
         int n = salary.length;
-        float totalSalary = 0;
+        int totalSalary = 0;
         for(int i=0; i<n; i++){
             totalSalary += salary[i];
         }
